Refactor prepare component
